fix(plugins): keep update modules below line budget
