+++ src/plugins/update.ts
@@ -1,13 +1,12 @@
 /** Updates installed plugins across npm, ClawHub, marketplace, Git, and bundled bridge sources. */
 export type {
-  PluginChannelSyncResult,
-  PluginChannelSyncSummary,
   PluginUpdateChannelFallback,
   PluginUpdateIntegrityDriftParams,
   PluginUpdateLogger,
   PluginUpdateOutcome,
   PluginUpdateSummary,
 } from "./update-source.js";
+export type { PluginChannelSyncResult, PluginChannelSyncSummary } from "./update-channel.js";
 
 export {
   isPluginInstallRecordUpdateSource,
